+++ Sources/RichTextKit/Coordinator/RichTextCoordinator+Subscriptions.swift
@@ -40,6 +40,8 @@ extension RichTextCoordinator {
                 self?.setColor(color, for: .foreground)
             case .backgroundColor(let color):
                 self?.setColor(color, for: .background)
+            case .underlineColor(let color):
+                self?.setColor(color, for: .underline)
             case .highlightedRange(let range):
                 self?.setHighlightedRange(to: range)
             case .highlightingStyle(let style):
@@ -163,13 +165,15 @@ internal extension RichTextCoordinator {
     }
 
     func setHighlightedRangeAppearance(for range: NSRange) {
-        highlightedRangeOriginalBackgroundColor = textView.richTextBackgroundColor(at: range) ?? .clear
-        highlightedRangeOriginalForegroundColor = textView.richTextForegroundColor(at: range) ?? .textColor
+        let back = textView.richTextColor(.background, at: range) ?? .clear
+        let fore = textView.richTextColor(.foreground, at: range) ?? .textColor
+        highlightedRangeOriginalBackgroundColor = back
+        highlightedRangeOriginalForegroundColor = fore
         let style = textView.highlightingStyle
         let background = ColorRepresentable(style.backgroundColor)
-        let text = ColorRepresentable(style.foregroundColor)
-        textView.setRichTextBackgroundColor(background, at: range)
-        textView.setRichTextForegroundColor(text, at: range)
+        let foreground = ColorRepresentable(style.foregroundColor)
+        textView.setRichTextColor(.background, to: background, at: range)
+        textView.setRichTextColor(.foreground, to: foreground, at: range)
     }
 
     func setIsEditing(to newValue: Bool) {