+++ Sources/RichTextKit/Coordinator/RichTextCoordinator.swift
@@ -148,10 +148,10 @@ public extension RichTextCoordinator {
         guard
             let range = richTextContext.highlightedRange,
             let background = highlightedRangeOriginalBackgroundColor,
-            let text = highlightedRangeOriginalForegroundColor
+            let foreground = highlightedRangeOriginalForegroundColor
         else { return }
-        textView.setRichTextBackgroundColor(background, at: range)
-        textView.setRichTextForegroundColor(text, at: range)
+        textView.setRichTextColor(.background, to: background, at: range)
+        textView.setRichTextColor(.foreground, to: foreground, at: range)
     }
 }
 
@@ -216,9 +216,14 @@ extension RichTextCoordinator {
             richTextContext.strokeColor = stroke
         }
 
-        let strike = textView.currentColor(.strikethrough)
-        if richTextContext.strikethroughColor != strike {
-            richTextContext.strikethroughColor = strike
+        let strikethrough = textView.currentColor(.strikethrough)
+        if richTextContext.strikethroughColor != strikethrough {
+            richTextContext.strikethroughColor = strikethrough
+        }
+
+        let underline = textView.currentColor(.underline)
+        if richTextContext.foregroundColor != underline {
+            richTextContext.foregroundColor = underline
         }
 
         let hasRange = textView.hasSelectedRange